Remove problematic cache infrastructure and consolidate distribution_cache
This commit completes the removal of the problematic AdaptiveCache and
MathFunctionCache infrastructure that was causing template instantiation
issues and cache invalidation problems. The ThreadSafeCacheManager remains
as the sole caching mechanism for distribution parameters.

Key changes:
- Removed AdaptiveCache and MathFunctionCache classes and their implementations
- Removed all test files for the deleted cache classes
- Removed cache-related diagnostic files that depended on AdaptiveCache
- Moved distribution_cache.h from include/cache/ to include/core/ as it's now
  a core component rather than optional cache infrastructure
- Removed cache_platform.h which provided vestigial cache configuration
- Removed the now-empty include/cache/ directory
- Updated all include paths to reference the new location of distribution_cache.h
- Cleaned up enhanced test files to remove AdaptiveCache references
- Updated CMakeLists.txt to remove references to deleted source and test files

The build succeeds with all major compilers (Clang, GCC, MSVC) in strict mode.
Some enhanced tests show SIMD performance expectations not being met, but this
is unrelated to the cache cleanup and likely due to CPU architecture or data
size factors.

The codebase is now cleaner and more maintainable with a single, well-tested
caching mechanism (ThreadSafeCacheManager) that provides thread-safe parameter
caching for distributions.
