+++ CMakeLists.txt
@@ -947,9 +947,7 @@ set(LIBSTATS_PLATFORM_SOURCES
 
 # Level 3: Advanced Infrastructure (Depends on Level 0-2)
 set(LIBSTATS_INFRASTRUCTURE_SOURCES
-    src/adaptive_cache.cpp            # Memory-aware adaptive caching
     src/distribution_cache.cpp        # Distribution-specific cache optimization
-    src/math_function_cache.cpp       # High-performance mathematical function caching
     src/benchmark.cpp                 # Performance benchmarking utilities
     src/performance_history.cpp       # Performance data collection and analysis
     src/performance_dispatcher.cpp    # Smart auto-dispatch system
@@ -1354,12 +1352,7 @@ if(LIBSTATS_BUILD_TESTS)
     # =======================================================================
     # Test advanced infrastructure components that depend on Level 0-2
     
-    # Performance and Caching Infrastructure
-    # adaptive_cache.h - Memory-aware adaptive caching
-    create_libstats_test(test_adaptive_cache tests/test_adaptive_cache.cpp)
-    
-    # math_function_cache.h - Mathematical function caching with precision rounding
-    create_libstats_gtest(test_math_function_cache tests/test_math_function_cache.cpp)
+    # Performance Infrastructure
     
     # parallel_execution.h - C++20 parallel execution with automatic fallback
     create_libstats_test(test_parallel_execution_integration tests/test_parallel_execution_integration.cpp)
@@ -1470,7 +1463,6 @@ if(LIBSTATS_BUILD_TESTS)
             # Level 0: Foundational tests (most critical)
             test_constants
             test_cpu_detection
-            test_adaptive_cache
             test_simd_integration_simple
             test_simd_integration
             test_simd_operations
