--- docs/cache_usage_analysis.md
@@ -1,120 +1,0 @@
-# Cache Usage Analysis
-
-## Overview
-This analysis examines which compilation units actually use the adaptive_cache.h and distribution_cache.h headers versus just including them through transitive dependencies.
-
-## Findings
-
-### Current Cache Strategy Landscape
-From the codebase analysis, the current `Strategy` enum includes:
-- `SCALAR` - Single element or very small batches
-- `SIMD_BATCH` - SIMD vectorized for medium batches  
-- `PARALLEL_SIMD` - Parallel + SIMD for large batches
-- `WORK_STEALING` - Dynamic load balancing for irregular workloads
-- `GPU_ACCELERATED` - GPU-accelerated execution (CPU fallback if GPU unavailable)
-
-**The `CACHE_AWARE` strategy has been removed**, which means the cache-aware dispatch code in distributions is now dead code.
-
-### Distribution Cache Header Inclusion
-All distribution headers include `distribution_platform_common.h` which transitively includes:
-- `adaptive_cache.h` 
-- `distribution_cache.h`
-
-**Distribution Headers including platform_common.h:**
-- `gaussian.h`
-- `uniform.h` 
-- `discrete.h`
-- `poisson.h`
-- `gamma.h`
-- `exponential.h`
-
-### Actual Cache Usage Analysis
-
-#### 1. Distribution Source Files
-**Exponential Distribution (`src/exponential.cpp`):** 
-- Contains **active cache usage** in CACHE_AWARE lambda functions (lines 1544-1598, 1721-1774)
-- Uses `cache.get()` and `cache.put()` calls for caching intermediate results
-- **Status:** Dead code - CACHE_AWARE strategy no longer exists
-
-**Gaussian Distribution (`src/gaussian.cpp`):**
-- **No direct cache usage** found via grep analysis
-- Only includes adaptive_cache.h transitively through platform_common.h
-- **Status:** Unnecessary inclusion
-
-**Other Distributions (uniform, discrete, poisson, gamma):**
-- **No direct cache usage** found in source files
-- Only transitive inclusion through platform_common.h
-- **Status:** Unnecessary inclusion
-
-#### 2. Base Infrastructure Files
-**DistributionBase (`include/core/distribution_base.h`):**
-- Directly includes both cache headers
-- Inherits from `ThreadSafeCacheManager` 
-- **Status:** Legitimate usage for base functionality
-
-**Distribution Cache (`src/distribution_cache.cpp`):**
-- Implements `DistributionCacheAdapter` and cache management
-- **Status:** Legitimate implementation file
-
-#### 3. Test and Diagnostic Files
-**Test Files:**
-- `test_adaptive_cache.cpp` - Direct cache testing (legitimate)
-- `test_cache_integration.cpp` - Cache integration testing (legitimate) 
-- Enhanced test templates include cache headers (legitimately for testing)
-
-**Diagnostic Files:**
-- Cache diagnostic tools legitimately use cache headers for analysis
-
-### Summary of Unnecessary Inclusions
-
-**Completely Unnecessary:**
-- `src/uniform.cpp` - includes adaptive_cache.h but no usage found
-- `src/discrete.cpp` - includes adaptive_cache.h but no usage found  
-- `src/gaussian.cpp` - no direct cache usage (centralized cache migration made this unnecessary)
-- `src/poisson.cpp` - no direct cache usage found
-- `src/gamma.cpp` - no direct cache usage found
-
-**Dead Code (CACHE_AWARE removal):**
-- `src/exponential.cpp` - Contains cache-aware lambda functions that are never called
-
-**Potentially Unnecessary Transitive Inclusion:**
-- `include/distributions/distribution_platform_common.h` - includes adaptive_cache.h for all distributions
-- Individual distribution headers inherit this through platform_common.h
-
-## Recommendations
-
-### 1. Remove Dead CACHE_AWARE Code
-Remove the cache-aware lambda functions from all distribution implementations since the CACHE_AWARE strategy no longer exists.
-
-### 2. Remove Unnecessary Direct Inclusions  
-Remove direct #include statements for adaptive_cache.h from distribution source files that don't use cache functionality:
-- `src/uniform.cpp`
-- `src/discrete.cpp`
-- `src/exponential.cpp` (after removing dead code)
-- `src/poisson.cpp`
-- `src/gamma.cpp`
-
-### 3. Reconsider Platform Common Inclusion
-Evaluate removing adaptive_cache.h from `distribution_platform_common.h` since:
-- Most distributions don't actually use cache functionality directly
-- The centralized cache system handles parameter caching differently
-- This would significantly reduce compilation dependencies
-
-### 4. Keep Base Infrastructure
-Maintain cache header inclusions in:
-- `distribution_base.h` - Base class legitimately uses cache infrastructure
-- `distribution_cache.cpp` - Implementation file for cache functionality
-- Test and diagnostic files - Legitimate usage for testing and analysis
-
-### 5. Alternative Approach
-Consider creating a separate `cache_platform_common.h` header that only distributions needing cache functionality would include, rather than including cache headers in the common platform header used by all distributions.
-
-## Impact Assessment
-
-**Compilation Time:** Removing unnecessary cache header inclusions could significantly improve compilation times by reducing template instantiations and header parsing.
-
-**Memory Usage:** Reduced memory footprint during compilation and potentially at runtime.
-
-**Maintenance:** Cleaner dependency graph makes the codebase easier to maintain and understand.
-
-**Functionality:** No impact on actual functionality since the removed inclusions aren't being used.
