+++ src/distribution_cache.cpp
@@ -1,4 +1,4 @@
-#include "cache/distribution_cache.h"
+#include "core/distribution_cache.h"
 #include <functional>
 
 namespace libstats {
