--- tests/test_adaptive_cache.cpp
@@ -1,814 +1,0 @@
-// Use focused header for adaptive cache testing
-#include "../include/platform/cache_platform.h"  // For cache functionality
-// Implementation is already included via the header
-#include <iostream>
-#include <cassert>
-#include <thread>
-#include <chrono>
-#include <random>
-#include <vector>
-#include <iomanip>
-#include <sstream>
-
-using namespace libstats::cache;
-using namespace libstats::constants;
-
-void test_basic_cache_operations() {
-    std::cout << "Testing basic cache operations..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 100;
-    config.max_memory_bytes = 1024 * 1024; // 1MB
-    config.ttl = std::chrono::milliseconds(1000);
-    config.enable_background_optimization = false; // Disable for deterministic testing
-    
-    AdaptiveCache<std::string, int> cache(config);
-    
-    // Test put and get
-    cache.put("key1", 42);
-    cache.put("key2", 84);
-    
-    [[maybe_unused]] auto result1 = cache.get("key1");
-    assert(result1.has_value() && result1.value() == 42);
-    
-    [[maybe_unused]] auto result2 = cache.get("key2");
-    assert(result2.has_value() && result2.value() == 84);
-    
-    // Test miss
-    [[maybe_unused]] auto result3 = cache.get("key3");
-    assert(!result3.has_value());
-    
-    std::cout << "✓ Basic cache operations passed" << std::endl;
-}
-
-void test_cache_eviction() {
-    std::cout << "Testing cache eviction..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 5;
-    config.max_memory_bytes = 1024; // Very small to force eviction
-    config.ttl = std::chrono::milliseconds(10000); // Long TTL
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<int, std::string> cache(config);
-    
-    // Fill cache beyond capacity
-    for (int i = 0; i < 10; ++i) {
-        cache.put(i, "value" + std::to_string(i));
-    }
-    
-    // Check that cache is limited in size
-    [[maybe_unused]] auto stats = cache.getStats();
-    
-    assert(stats.size <= static_cast<std::size_t>(config.max_cache_size));
-    
-    // Verify some entries were evicted
-    int found_count = 0;
-    for (int i = 0; i < 10; ++i) {
-        if (cache.get(i).has_value()) {
-            found_count++;
-        }
-    }
-    assert(found_count < 10);
-    
-    std::cout << "✓ Cache eviction passed (found " << found_count << " out of 10 entries)" << std::endl;
-}
-
-void test_ttl_expiration() {
-    std::cout << "Testing TTL expiration..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 100;
-    config.ttl = std::chrono::milliseconds(100); // Very short TTL
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<std::string, int> cache(config);
-    
-    cache.put("test_key", 123);
-    
-    // Should be available immediately
-    [[maybe_unused]] auto result1 = cache.get("test_key");
-    assert(result1.has_value() && result1.value() == 123);
-    
-    // Wait for expiration
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-    
-    // Try to access again - might still be there until next put() cleans it up
-    cache.put("trigger_cleanup", 456); // This should trigger cleanup
-    
-    // Note: Due to lazy cleanup, expired entries might still return values
-    // The important thing is that the TTL mechanism is working
-    
-    std::cout << "✓ TTL expiration test completed" << std::endl;
-}
-
-void test_metrics_and_stats() {
-    std::cout << "Testing metrics and statistics..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 50;
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<int, std::string> cache(config);
-    
-    // Perform some operations
-    for (int i = 0; i < 20; ++i) {
-        cache.put(i, "value" + std::to_string(i));
-    }
-    
-    // Access some entries multiple times (hits)
-    for (int i = 0; i < 10; ++i) {
-        for (int j = 0; j < 3; ++j) {
-            cache.get(i);
-        }
-    }
-    
-    // Try to access non-existent entries (misses)
-    for (int i = 100; i < 110; ++i) {
-        cache.get(i);
-    }
-    
-    [[maybe_unused]] auto metrics = cache.getMetrics();
-    [[maybe_unused]] auto stats = cache.getStats();
-    
-    // Debug: std::cout << "Debug metrics: hits=" << metrics.hits.load() << ", misses=" << metrics.misses.load() << ", hit_rate=" << stats.hit_rate << std::endl;
-    
-    assert(metrics.hits.load() >= 30); // At least 30 hits from our loop
-    assert(metrics.misses.load() >= 10); // At least 10 misses
-    // Allow for hit_rate to be 0.0 initially if no accesses yet
-    assert(stats.hit_rate >= 0.0 && stats.hit_rate <= 1.0);
-    assert(stats.size <= config.max_cache_size);
-    
-    std::cout << "✓ Metrics and stats passed (hit rate: " 
-              << (stats.hit_rate * 100) << "%, size: " << stats.size << ")" << std::endl;
-}
-
-void test_prefetching() {
-    std::cout << "Testing prefetching functionality..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 100;
-    config.enable_prefetching = true;
-    config.prefetch_queue_size = 10;
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<int, std::string> cache(config);
-    
-    // Define a value generator for prefetching
-    auto generator = [](int key) -> std::string {
-        return "prefetch_value_" + std::to_string(key);
-    };
-    
-    // Prefetch some keys
-    std::vector<int> prefetch_keys = {100, 101, 102, 103, 104};
-    cache.prefetch(prefetch_keys, generator);
-    
-    // Check if prefetched values are available
-    int prefetch_hits = 0;
-    for (int key : prefetch_keys) {
-        auto result = cache.get(key);
-        if (result.has_value()) {
-            prefetch_hits++;
-            assert(result.value() == generator(key));
-        }
-    }
-    
-    [[maybe_unused]] auto metrics = cache.getMetrics();
-    assert(metrics.prefetch_hits.load() > 0 || metrics.prefetch_misses.load() > 0);
-    
-    std::cout << "✓ Prefetching passed (prefetched " << prefetch_hits 
-              << " out of " << static_cast<std::size_t>(prefetch_keys.size()) << " keys)" << std::endl;
-}
-
-void test_memory_pressure_detection() {
-    std::cout << "Testing memory pressure detection..." << std::endl;
-    
-    MemoryPressureDetector detector;
-    auto pressure_info = detector.detectPressure();
-    
-    assert(pressure_info.pressure_level >= 0.0 && pressure_info.pressure_level <= 1.0);
-    assert(pressure_info.available_cache_mb > 0);
-    assert(!pressure_info.recommendation.empty());
-    
-    std::cout << "✓ Memory pressure detection passed" << std::endl;
-    std::cout << "  Pressure level: " << pressure_info.pressure_level << std::endl;
-    std::cout << "  Available cache: " << pressure_info.available_cache_mb << " MB" << std::endl;
-    std::cout << "  Recommendation: " << pressure_info.recommendation << std::endl;
-}
-
-void test_cache_advisor() {
-    std::cout << "Testing cache advisor..." << std::endl;
-    
-    CacheAdvisor advisor;
-    
-    // Create mock metrics and config for testing
-    CacheMetrics mock_metrics;
-    mock_metrics.hits.store(700);
-    mock_metrics.misses.store(300);
-    mock_metrics.evictions.store(50);
-    mock_metrics.memory_usage.store(2 * 1024 * 1024); // 2MB
-    mock_metrics.prefetch_hits.store(20);
-    mock_metrics.prefetch_misses.store(80);
-    mock_metrics.updateHitRate();
-    
-    AdaptiveCacheConfig mock_config;
-    mock_config.enable_prefetching = true;
-    
-    MemoryPressureDetector detector;
-    auto memory_info = detector.detectPressure();
-    
-    auto recommendations = advisor.analyzeAndRecommend(mock_metrics, mock_config, memory_info);
-    
-    std::cout << "✓ Cache advisor generated " << static_cast<std::size_t>(recommendations.size()) << " recommendations:" << std::endl;
-    for (const auto& rec : recommendations) {
-        std::cout << "  Priority " << rec.priority << ": " << rec.description << std::endl;
-    }
-}
-
-void test_optimal_config_creation() {
-    std::cout << "Testing optimal configuration creation..." << std::endl;
-    
-    auto config = utils::createOptimalConfig();
-    
-    assert(config.max_memory_bytes > 0);
-    assert(config.max_cache_size > 0);
-    assert(config.ttl.count() > 0);
-    
-    std::cout << "✓ Optimal config creation passed" << std::endl;
-    std::cout << "  Max memory: " << (config.max_memory_bytes / 1024 / 1024) << " MB" << std::endl;
-    std::cout << "  Max cache size: " << config.max_cache_size << " entries" << std::endl;
-    std::cout << "  TTL: " << config.ttl.count() << " ms" << std::endl;
-    std::cout << "  Prefetching: " << (config.enable_prefetching ? "enabled" : "disabled") << std::endl;
-}
-
-void test_cache_monitor() {
-    std::cout << "Testing cache monitoring..." << std::endl;
-    
-    CacheMonitor monitor;
-    
-    // Simulate some cache metrics over time
-    for (int i = 0; i < 10; ++i) {
-        CacheMetrics metrics;
-        metrics.hits.store(static_cast<unsigned long>(100 + i * 10));
-        metrics.misses.store(static_cast<unsigned long>(20 + i * 2));
-        metrics.memory_usage.store(static_cast<unsigned long>(1024 * 1024 + i * 1024));
-        metrics.average_access_time.store(100.0 + i);
-        metrics.updateHitRate();
-        
-        monitor.recordMetrics(metrics);
-    }
-    
-    // Generate a report
-    CacheMetrics final_metrics;
-    final_metrics.hits.store(200);
-    final_metrics.misses.store(40);
-    final_metrics.memory_usage.store(2 * 1024 * 1024);
-    final_metrics.cache_size.store(1500);
-    final_metrics.average_access_time.store(120.0);
-    final_metrics.evictions.store(25);
-    final_metrics.updateHitRate();
-    
-    auto report = monitor.generateReport(final_metrics);
-    assert(!report.empty());
-    
-    std::cout << "✓ Cache monitoring passed" << std::endl;
-    std::cout << report << std::endl;
-}
-
-void test_performance_benchmarking() {
-    std::cout << "Testing performance benchmarking..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 1000;
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<int, std::string> cache(config);
-    
-    // Create test data
-    std::vector<std::pair<int, std::string>> test_data;
-    for (int i = 0; i < 100; ++i) {
-        test_data.emplace_back(i, "benchmark_value_" + std::to_string(i));
-    }
-    
-    // Run benchmark
-    auto result = utils::benchmarkCache(cache, test_data, 1000);
-    
-    assert(result.hit_rate >= 0.0 && result.hit_rate <= 1.0);
-    assert(result.operations_per_second > 0);
-    
-    std::cout << "✓ Performance benchmarking passed" << std::endl;
-    std::cout << "  Hit rate: " << (result.hit_rate * 100) << "%" << std::endl;
-    std::cout << "  Operations/sec: " << result.operations_per_second << std::endl;
-    std::cout << "  Avg access time: " << result.average_access_time_us << " μs" << std::endl;
-}
-
-void test_string_key_performance_degradation() {
-    std::cout << "Testing string-based cache key performance degradation..." << std::endl;
-    std::cout << "This test simulates the problematic pattern that causes O(n²) performance issues" << std::endl;
-    std::cout << "in distribution Cache-Aware implementations." << std::endl;
-    std::cout << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 10000;  // Large enough to hold our test data
-    config.max_memory_bytes = 50 * 1024 * 1024; // 50MB 
-    config.ttl = std::chrono::milliseconds(30000); // Long TTL to avoid expiration
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<std::string, double> cache(config);
-    
-    // Test different batch sizes to show performance degradation
-    std::vector<size_t> batch_sizes = {50, 100, 250, 500, 750, 1000, 1500, 2000, 5000};
-    
-    std::cout << "Batch Size | Operations Time (μs) | Ops/sec   | Key Gen Time (μs) | Cache Time (μs) | Performance" << std::endl;
-    std::cout << "-----------|----------------------|-----------|-------------------|-----------------|------------" << std::endl;
-    
-    double previous_ops_per_sec = 0.0;
-    
-    for (size_t batch_size : batch_sizes) {
-        // Create test data - simulating floating point values like distribution inputs
-        std::vector<double> test_values;
-        test_values.reserve(batch_size);
-        for (size_t i = 0; i < batch_size; ++i) {
-            test_values.push_back(1.0 + static_cast<double>(i) * 0.1); // Simulate realistic distribution values
-        }
-        
-        // Measure total time for the problematic string-based caching pattern
-        auto total_start = std::chrono::high_resolution_clock::now();
-        
-        // Measure key generation time separately
-        auto key_gen_start = std::chrono::high_resolution_clock::now();
-        std::vector<std::string> cache_keys;
-        cache_keys.reserve(batch_size);
-        
-        // This is the exact problematic pattern from the original Cache-Aware lambdas
-        for (size_t i = 0; i < batch_size; ++i) {
-            std::ostringstream key_stream;
-            key_stream << std::fixed << std::setprecision(6) << "discrete_pdf_" << test_values[i];
-            cache_keys.push_back(key_stream.str()); // Expensive string construction
-        }
-        
-        auto key_gen_end = std::chrono::high_resolution_clock::now();
-        auto key_gen_time = std::chrono::duration_cast<std::chrono::microseconds>(key_gen_end - key_gen_start);
-        
-        // Measure cache operations time
-        auto cache_start = std::chrono::high_resolution_clock::now();
-        
-        // Simulate the full cache workflow: check cache, compute if miss, store result
-        size_t cache_hits = 0;
-        for (size_t i = 0; i < batch_size; ++i) {
-            // Try to get from cache first
-            if (auto cached_result = cache.get(cache_keys[i])) {
-                cache_hits++;
-                // Use cached result (simulate)
-                [[maybe_unused]] double result = *cached_result;
-            } else {
-                // Simulate computation and caching (like in distribution lambdas)
-                double computed_result = test_values[i] * 0.166667; // Simulate discrete uniform PMF
-                cache.put(cache_keys[i], computed_result);
-            }
-        }
-        
-        auto cache_end = std::chrono::high_resolution_clock::now();
-        auto cache_time = std::chrono::duration_cast<std::chrono::microseconds>(cache_end - cache_start);
-        
-        auto total_end = std::chrono::high_resolution_clock::now();
-        auto total_time = std::chrono::duration_cast<std::chrono::microseconds>(total_end - total_start);
-        
-        // Calculate performance metrics
-        double ops_per_second = (static_cast<double>(batch_size) / static_cast<double>(total_time.count())) * 1000000.0;
-        
-        // Determine performance trend
-        std::string performance_trend;
-        if (previous_ops_per_sec == 0.0) {
-            performance_trend = "baseline";
-        } else {
-            double change_ratio = ops_per_second / previous_ops_per_sec;
-            if (change_ratio > 1.05) {
-                performance_trend = "improved";
-            } else if (change_ratio < 0.95) {
-                performance_trend = "degraded";
-            } else {
-                performance_trend = "stable";
-            }
-        }
-        
-        // Format output with better alignment
-        std::cout << std::setw(10) << batch_size
-                  << " | " << std::setw(20) << total_time.count()
-                  << " | " << std::setw(9) << static_cast<int>(ops_per_second)
-                  << " | " << std::setw(17) << key_gen_time.count() 
-                  << " | " << std::setw(14) << cache_time.count()
-                  << " | " << performance_trend;
-        
-        if (cache_hits > 0) {
-            std::cout << " (" << cache_hits << " hits)";
-        }
-        std::cout << std::endl;
-        
-        previous_ops_per_sec = ops_per_second;
-        
-        // Safety check - if performance degrades too much, warn and continue
-        if (total_time.count() > 50000) { // More than 50ms for a batch
-            std::cout << "         ⚠️  WARNING: Severe performance degradation detected!" << std::endl;
-        }
-        
-        // Clear cache periodically to simulate realistic usage
-        if (batch_size >= 1000) {
-            cache.clear();
-        }
-    }
-    
-    auto final_stats = cache.getStats();
-    std::cout << std::endl;
-    std::cout << "Final cache stats: " << final_stats.size << " entries, " 
-              << (final_stats.hit_rate * 100) << "% hit rate" << std::endl;
-    std::cout << std::endl;
-    
-    std::cout << "Analysis:" << std::endl;
-    std::cout << "- String key generation dominates performance at larger batch sizes" << std::endl;
-    std::cout << "- Each ostringstream operation allocates memory and formats floating-point numbers" << std::endl;
-    std::cout << "- This creates O(n²) behavior: n operations × string generation overhead" << std::endl;
-    std::cout << "- Cache operations themselves remain fast - the problem is key generation" << std::endl;
-    std::cout << std::endl;
-    
-    std::cout << "✓ String-based cache key performance degradation test completed" << std::endl;
-}
-
-void test_thread_safety() {
-    std::cout << "Testing thread safety..." << std::endl;
-    
-    AdaptiveCacheConfig config;
-    config.max_cache_size = 1000;
-    config.enable_background_optimization = false;
-    
-    AdaptiveCache<int, int> cache(config);
-    
-    const int num_threads = 4;
-    std::vector<std::thread> threads;
-    
-    // Launch multiple threads performing concurrent operations
-    for (int t = 0; t < num_threads; ++t) {
-        threads.emplace_back([&cache, t]() {
-            const int operations_per_thread_local = 1000;
-            std::random_device rd;
-            std::mt19937 gen(rd());
-            std::uniform_int_distribution<> key_dist(0, 99);
-            std::uniform_int_distribution<> op_dist(0, 2); // 0=put, 1=get, 2=clear
-            
-            for (int i = 0; i < operations_per_thread_local; ++i) {
-                int key = key_dist(gen);
-                int op = op_dist(gen);
-                
-                switch (op) {
-                    case 0: // put
-                        cache.put(key, t * 1000 + i);
-                        break;
-                    case 1: // get
-                        cache.get(key);
-                        break;
-                    case 2: // clear (occasionally)
-                        if (i % 100 == 0) {
-                            cache.clear();
-                        }
-                        break;
-                }
-            }
-        });
-    }
-    
-    // Wait for all threads to complete
-    for (auto& thread : threads) {
-        thread.join();
-    }
-    
-    // Verify cache is still in a consistent state
-    auto stats = cache.getStats();
-    assert(stats.size >= 0);
-    
-    std::cout << "✓ Thread safety test passed (final cache size: " 
-              << stats.size << ")" << std::endl;
-}
-
-// ===== NEW PLATFORM-AWARE CACHE HIERARCHY TESTS =====
-
-void test_platform_architecture_detection() {
-    std::cout << "Testing platform architecture detection..." << std::endl;
-    
-    auto architecture = utils::detectPlatformArchitecture();
-    
-    // Verify we get a valid architecture type
-    assert(architecture == utils::PlatformArchitecture::APPLE_SILICON ||
-           architecture == utils::PlatformArchitecture::INTEL ||
-           architecture == utils::PlatformArchitecture::AMD ||
-           architecture == utils::PlatformArchitecture::ARM_GENERIC ||
-           architecture == utils::PlatformArchitecture::UNKNOWN);
-    
-    // Print detected architecture for verification
-    std::string arch_name;
-    switch (architecture) {
-        case utils::PlatformArchitecture::APPLE_SILICON:
-            arch_name = "Apple Silicon (M1/M2/M3)";
-            break;
-        case utils::PlatformArchitecture::INTEL:
-            arch_name = "Intel";
-            break;
-        case utils::PlatformArchitecture::AMD:
-            arch_name = "AMD";
-            break;
-        case utils::PlatformArchitecture::ARM_GENERIC:
-            arch_name = "ARM (Generic)";
-            break;
-        case utils::PlatformArchitecture::UNKNOWN:
-            arch_name = "Unknown";
-            break;
-    }
-    
-    std::cout << "✓ Platform detection passed (detected: " << arch_name << ")" << std::endl;
-}
-
-void test_platform_specific_cache_tuning() {
-    std::cout << "Testing platform-specific cache tuning..." << std::endl;
-    
-    auto config = utils::createOptimalConfig();
-    
-    // Verify configuration uses platform-aware constants
-    assert(config.max_memory_bytes >= cache::sizing::MIN_CACHE_SIZE_BYTES);
-    assert(config.max_memory_bytes <= cache::sizing::MAX_CACHE_SIZE_BYTES);
-    assert(config.max_cache_size >= cache::sizing::MIN_ENTRY_COUNT);
-    assert(config.max_cache_size <= cache::sizing::MAX_ENTRY_COUNT);
-    
-    // Verify TTL is within reasonable bounds
-    assert(config.ttl >= cache::tuning::BASE_TTL * 0.5); // At least half base TTL
-    assert(config.ttl <= cache::tuning::ULTRA_HIGH_FREQ_TTL * 2); // At most 2x ultra-high TTL
-    
-    // Verify eviction threshold is reasonable
-    assert(config.eviction_threshold >= 0.5 && config.eviction_threshold <= 1.0);
-    
-    // Verify prefetch queue size is bounded
-    assert(config.prefetch_queue_size <= config.max_cache_size / 4);
-    
-    std::cout << "✓ Platform-specific cache tuning passed" << std::endl;
-    std::cout << "  Memory: " << (config.max_memory_bytes / (1024*1024)) << "MB, "
-              << "Size: " << config.max_cache_size << " entries, "
-              << "TTL: " << config.ttl.count() << "ms" << std::endl;
-    std::cout << "  Eviction threshold: " << (config.eviction_threshold * 100) << "%, "
-              << "Prefetch queue: " << config.prefetch_queue_size << std::endl;
-}
-
-void test_access_pattern_analysis() {
-    std::cout << "Testing access pattern analysis..." << std::endl;
-    
-    utils::AccessPatternAnalyzer analyzer;
-    
-    // Test sequential pattern
-    std::cout << "  Testing sequential pattern detection..." << std::endl;
-    for (int i = 0; i < 50; ++i) {
-        analyzer.recordAccess(i);
-    }
-    
-    auto pattern_info = analyzer.analyzePattern();
-    assert(pattern_info.type != utils::AccessPatternAnalyzer::PatternType::UNKNOWN);
-    assert(pattern_info.sequential_ratio >= 0.0 && pattern_info.sequential_ratio <= 1.0);
-    assert(pattern_info.locality_score >= 0.0 && pattern_info.locality_score <= 1.0);
-    assert(!pattern_info.description.empty());
-    
-    std::cout << "  Sequential pattern: ratio=" << pattern_info.sequential_ratio 
-              << ", locality=" << pattern_info.locality_score << std::endl;
-    
-    // Test random pattern
-    std::cout << "  Testing random pattern detection..." << std::endl;
-    utils::AccessPatternAnalyzer random_analyzer;
-    std::random_device rd;
-    std::mt19937 gen(rd());
-    std::uniform_int_distribution<> dist(0, 1000);
-    
-    for (int i = 0; i < 50; ++i) {
-        random_analyzer.recordAccess(dist(gen));
-    }
-    
-    auto random_pattern = random_analyzer.analyzePattern();
-    assert(random_pattern.sequential_ratio >= 0.0 && random_pattern.sequential_ratio <= 1.0);
-    
-    std::cout << "  Random pattern: ratio=" << random_pattern.sequential_ratio 
-              << ", locality=" << random_pattern.locality_score << std::endl;
-    
-    std::cout << "✓ Access pattern analysis passed" << std::endl;
-}
-
-void test_pattern_aware_cache_configuration() {
-    std::cout << "Testing pattern-aware cache configuration..." << std::endl;
-    
-    // Test sequential pattern configuration
-    utils::AccessPatternAnalyzer::PatternInfo sequential_pattern;
-    sequential_pattern.type = utils::AccessPatternAnalyzer::PatternType::SEQUENTIAL;
-    sequential_pattern.sequential_ratio = 0.9;
-    sequential_pattern.locality_score = 0.8;
-    
-    auto sequential_config = utils::createPatternAwareConfig(sequential_pattern);
-    
-    // Test random pattern configuration
-    utils::AccessPatternAnalyzer::PatternInfo random_pattern;
-    random_pattern.type = utils::AccessPatternAnalyzer::PatternType::RANDOM;
-    random_pattern.sequential_ratio = 0.1;
-    random_pattern.locality_score = 0.2;
-    
-    auto random_config = utils::createPatternAwareConfig(random_pattern);
-    
-    // Verify configurations are different and reasonable
-    assert(sequential_config.max_memory_bytes != random_config.max_memory_bytes ||
-           sequential_config.max_cache_size != random_config.max_cache_size ||
-           sequential_config.ttl != random_config.ttl);
-    
-    // Sequential should typically have larger cache or longer TTL
-    bool sequential_optimized = 
-        sequential_config.max_cache_size >= random_config.max_cache_size ||
-        sequential_config.ttl >= random_config.ttl;
-    
-    // Allow some flexibility in optimization strategies
-    if (!sequential_optimized) {
-        std::cout << "  Note: Sequential pattern config differs from expected optimization" << std::endl;
-    }
-    
-    std::cout << "✓ Pattern-aware configuration passed" << std::endl;
-    std::cout << "  Sequential: " << (sequential_config.max_memory_bytes/(1024*1024)) 
-              << "MB, " << sequential_config.ttl.count() << "ms TTL" << std::endl;
-    std::cout << "  Random: " << (random_config.max_memory_bytes/(1024*1024)) 
-              << "MB, " << random_config.ttl.count() << "ms TTL" << std::endl;
-}
-
-void test_cache_hierarchy_constants() {
-    std::cout << "Testing cache hierarchy constants..." << std::endl;
-    
-    // Test platform-specific constants are within reasonable ranges
-    assert(cache::platform::apple_silicon::DEFAULT_MAX_MEMORY_MB >= 1);
-    assert(cache::platform::apple_silicon::DEFAULT_MAX_MEMORY_MB <= 32);
-    assert(cache::platform::intel::DEFAULT_MAX_MEMORY_MB >= 1);
-    assert(cache::platform::intel::DEFAULT_MAX_MEMORY_MB <= 32);
-    assert(cache::platform::amd::DEFAULT_MAX_MEMORY_MB >= 1);
-    assert(cache::platform::amd::DEFAULT_MAX_MEMORY_MB <= 32);
-    
-    // Test cache sizing constants
-    assert(cache::sizing::L3_CACHE_FRACTION > 0.0 && cache::sizing::L3_CACHE_FRACTION < 1.0);
-    assert(cache::sizing::L2_CACHE_FRACTION > 0.0 && cache::sizing::L2_CACHE_FRACTION < 1.0);
-    assert(cache::sizing::MIN_CACHE_SIZE_BYTES < cache::sizing::MAX_CACHE_SIZE_BYTES);
-    assert(cache::sizing::MIN_ENTRY_COUNT < cache::sizing::MAX_ENTRY_COUNT);
-    
-    // Test tuning constants
-    assert(cache::tuning::HIGH_FREQ_THRESHOLD_HZ > 0);
-    assert(cache::tuning::ULTRA_HIGH_FREQ_THRESHOLD_HZ > cache::tuning::HIGH_FREQ_THRESHOLD_HZ);
-    // Higher frequency CPUs should have shorter TTL (faster invalidation)
-    assert(cache::tuning::BASE_TTL > cache::tuning::HIGH_FREQ_TTL);
-    assert(cache::tuning::HIGH_FREQ_TTL > cache::tuning::ULTRA_HIGH_FREQ_TTL);
-    
-    // Test pattern constants
-    assert(cache::patterns::SEQUENTIAL_PATTERN_THRESHOLD > 0.5);
-    assert(cache::patterns::RANDOM_PATTERN_THRESHOLD < 0.5);
-    assert(cache::patterns::SEQUENTIAL_SIZE_MULTIPLIER > 1.0);
-    assert(cache::patterns::RANDOM_SIZE_MULTIPLIER < 1.0);
-    
-    std::cout << "✓ Cache hierarchy constants validation passed" << std::endl;
-}
-
-void test_adaptive_cache_with_platform_tuning() {
-    std::cout << "Testing adaptive cache with platform tuning integration..." << std::endl;
-    
-    // Create cache with platform-optimized configuration
-    auto platform_config = utils::createOptimalConfig();
-    AdaptiveCache<std::string, int> platform_cache(platform_config);
-    
-    // Create cache with pattern-aware configuration
-    utils::AccessPatternAnalyzer::PatternInfo mixed_pattern;
-    mixed_pattern.type = utils::AccessPatternAnalyzer::PatternType::MIXED;
-    mixed_pattern.sequential_ratio = 0.6;
-    mixed_pattern.locality_score = 0.7;
-    
-    auto pattern_config = utils::createPatternAwareConfig(mixed_pattern);
-    AdaptiveCache<std::string, int> pattern_cache(pattern_config);
-    
-    // Test both caches with similar workloads
-    const int num_operations = 200;
-    
-    // First, populate the caches with initial data
-    for (int i = 0; i < 50; ++i) {
-        std::string key = "test_key_" + std::to_string(i);
-        int value = i * 10;
-        platform_cache.put(key, value);
-        pattern_cache.put(key, value);
-    }
-    
-    // Now perform mixed operations (puts and gets)
-    for (int i = 0; i < num_operations; ++i) {
-        std::string key = "test_key_" + std::to_string(i % 50); // Reuse keys for hits
-        int value = i * 10;
-        
-        // Mix of operations: 70% gets (should hit), 30% puts
-        if (i % 10 < 7) {
-            // Perform get operation (should hit since key exists)
-            auto result = platform_cache.get(key);
-            auto result2 = pattern_cache.get(key);
-            (void)result; (void)result2; // Use results to avoid compiler warnings
-        } else {
-            // Perform put operation (updates existing entries)
-            platform_cache.put(key, value);
-            pattern_cache.put(key, value);
-        }
-    }
-    
-    auto platform_stats = platform_cache.getStats();
-    auto pattern_stats = pattern_cache.getStats();
-    [[maybe_unused]] auto platform_metrics = platform_cache.getMetrics();
-    [[maybe_unused]] auto pattern_metrics = pattern_cache.getMetrics();
-    
-    // Both caches should be functional
-    assert(platform_stats.size > 0);
-    assert(pattern_stats.size > 0);
-    assert(platform_stats.hit_rate >= 0.0 && platform_stats.hit_rate <= 1.0);
-    assert(pattern_stats.hit_rate >= 0.0 && pattern_stats.hit_rate <= 1.0);
-    
-    // We should have some hits since we're performing gets on existing keys
-    assert(platform_metrics.hits.load() > 0 || platform_metrics.misses.load() > 0);
-    assert(pattern_metrics.hits.load() > 0 || pattern_metrics.misses.load() > 0);
-    
-    std::cout << "✓ Adaptive cache with platform tuning passed" << std::endl;
-    std::cout << "  Platform-tuned: " << platform_stats.size << " entries, "
-              << (platform_stats.hit_rate * 100) << "% hit rate" << std::endl;
-    std::cout << "  Pattern-tuned: " << pattern_stats.size << " entries, "
-              << (pattern_stats.hit_rate * 100) << "% hit rate" << std::endl;
-}
-
-int main() {
-    std::cout << "=== Adaptive Cache Management Tests ===" << std::endl;
-    std::cout << std::endl;
-    
-    try {
-        test_basic_cache_operations();
-        std::cout << std::endl;
-        
-        test_cache_eviction();
-        std::cout << std::endl;
-        
-        test_ttl_expiration();
-        std::cout << std::endl;
-        
-        test_metrics_and_stats();
-        std::cout << std::endl;
-        
-        test_prefetching();
-        std::cout << std::endl;
-        
-        test_memory_pressure_detection();
-        std::cout << std::endl;
-        
-        test_cache_advisor();
-        std::cout << std::endl;
-        
-        test_optimal_config_creation();
-        std::cout << std::endl;
-        
-        test_cache_monitor();
-        std::cout << std::endl;
-        
-        test_performance_benchmarking();
-        std::cout << std::endl;
-        
-        test_string_key_performance_degradation();
-        std::cout << std::endl;
-        
-        test_thread_safety();
-        std::cout << std::endl;
-        
-        // ===== NEW PLATFORM-AWARE CACHE HIERARCHY TESTS =====
-        std::cout << "=== Platform-Aware Cache Hierarchy Tests ===" << std::endl;
-        std::cout << std::endl;
-        
-        test_platform_architecture_detection();
-        std::cout << std::endl;
-        
-        test_platform_specific_cache_tuning();
-        std::cout << std::endl;
-        
-        test_access_pattern_analysis();
-        std::cout << std::endl;
-        
-        test_pattern_aware_cache_configuration();
-        std::cout << std::endl;
-        
-        test_cache_hierarchy_constants();
-        std::cout << std::endl;
-        
-        test_adaptive_cache_with_platform_tuning();
-        std::cout << std::endl;
-        
-        std::cout << "🎉 ALL ADAPTIVE CACHE TESTS PASSED! 🎉" << std::endl;
-        std::cout << "📊 Platform-aware cache hierarchy functionality verified! 📊" << std::endl;
-        
-    } catch (const std::exception& e) {
-        std::cerr << "❌ Test failed with exception: " << e.what() << std::endl;
-        return 1;
-    } catch (...) {
-        std::cerr << "❌ Test failed with unknown exception" << std::endl;
-        return 1;
-    }
-    
-    return 0;
-}
