--- tests/test_math_function_cache.cpp
@@ -1,447 +1,0 @@
-#include <gtest/gtest.h>
-#include "../include/cache/math_function_cache.h"
-#include <chrono>
-#include <cmath>
-#include <vector>
-
-using namespace libstats::cache;
-
-class MathFunctionCacheTest : public ::testing::Test {
-protected:
-    void SetUp() override {
-        // Initialize cache with test configuration
-        MathFunctionCacheConfig config;
-        config.gamma_cache_size = 512;
-        config.erf_cache_size = 256;
-        config.beta_cache_size = 256;
-        config.log_cache_size = 128;
-        config.enable_statistics = true;
-        config.gamma_precision = 0.001;
-        config.erf_precision = 0.0001;
-        config.beta_precision = 0.001;
-        config.log_precision = 0.0001;
-        
-        MathFunctionCache::initialize(config);
-    }
-    
-    void TearDown() override {
-        // Clear all caches for clean state
-        MathFunctionCache::clearAll();
-    }
-};
-
-TEST_F(MathFunctionCacheTest, GammaFunctionCaching) {
-    // Test basic gamma function caching
-    double x = 2.5;
-    
-    // First call should be a cache miss
-    auto start1 = std::chrono::high_resolution_clock::now();
-    double result1 = MathFunctionCache::getCachedGamma(x);
-    auto end1 = std::chrono::high_resolution_clock::now();
-    auto time1 = std::chrono::duration_cast<std::chrono::nanoseconds>(end1 - start1).count();
-    
-    // Second call should be a cache hit (faster)
-    auto start2 = std::chrono::high_resolution_clock::now();
-    double result2 = MathFunctionCache::getCachedGamma(x);
-    auto end2 = std::chrono::high_resolution_clock::now();
-    auto time2 = std::chrono::duration_cast<std::chrono::nanoseconds>(end2 - start2).count();
-    
-    // Results should be identical
-    EXPECT_EQ(result1, result2);
-    
-    // Verify result is correct (gamma(2.5) = 1.5 * gamma(1.5) = 1.5 * sqrt(pi)/2)
-    double expected = std::tgamma(x);
-    EXPECT_NEAR(result1, expected, 1e-10);
-    
-    // Second call should be faster (cache hit)
-    EXPECT_LT(time2, time1);
-    
-    // Check cache statistics
-    auto stats = MathFunctionCache::getStats();
-    EXPECT_GE(stats.gamma_hits.load(), 1);
-    EXPECT_GE(stats.gamma_misses.load(), 1);
-}
-
-TEST_F(MathFunctionCacheTest, ErrorFunctionCaching) {
-    // Test erf and erfc function caching
-    double x = 1.5;
-    
-    // Test erf caching
-    double erf1 = MathFunctionCache::getCachedErf(x);
-    double erf2 = MathFunctionCache::getCachedErf(x);
-    EXPECT_EQ(erf1, erf2);
-    EXPECT_NEAR(erf1, std::erf(x), 1e-10);
-    
-    // Test erfc caching
-    double erfc1 = MathFunctionCache::getCachedErfc(x);
-    double erfc2 = MathFunctionCache::getCachedErfc(x);
-    EXPECT_EQ(erfc1, erfc2);
-    EXPECT_NEAR(erfc1, std::erfc(x), 1e-10);
-    
-    // Verify mathematical relationship: erf(x) + erfc(x) = 1
-    EXPECT_NEAR(erf1 + erfc1, 1.0, 1e-10);
-}
-
-TEST_F(MathFunctionCacheTest, BetaFunctionCaching) {
-    // Test beta function caching
-    double a = 2.0, b = 3.0;
-    
-    double beta1 = MathFunctionCache::getCachedBeta(a, b);
-    double beta2 = MathFunctionCache::getCachedBeta(a, b);
-    EXPECT_EQ(beta1, beta2);
-    
-    // Verify result using beta(a,b) = gamma(a)*gamma(b)/gamma(a+b)
-    double expected = std::tgamma(a) * std::tgamma(b) / std::tgamma(a + b);
-    EXPECT_NEAR(beta1, expected, 1e-10);
-    
-    // Test symmetry: beta(a,b) = beta(b,a)
-    double beta_symmetric = MathFunctionCache::getCachedBeta(b, a);
-    EXPECT_NEAR(beta1, beta_symmetric, 1e-10);
-}
-
-TEST_F(MathFunctionCacheTest, LogarithmFunctionCaching) {
-    // Test natural and base-10 logarithm caching
-    double x = 10.0;
-    
-    // Test natural log
-    double log1 = MathFunctionCache::getCachedLog(x);
-    double log2 = MathFunctionCache::getCachedLog(x);
-    EXPECT_EQ(log1, log2);
-    EXPECT_NEAR(log1, std::log(x), 1e-10);
-    
-    // Test base-10 log
-    double log10_1 = MathFunctionCache::getCachedLog10(x);
-    double log10_2 = MathFunctionCache::getCachedLog10(x);
-    EXPECT_EQ(log10_1, log10_2);
-    EXPECT_NEAR(log10_1, std::log10(x), 1e-10);
-    
-    // Verify relationship: log10(x) = ln(x) / ln(10)
-    EXPECT_NEAR(log10_1, log1 / std::log(10.0), 1e-10);
-}
-
-TEST_F(MathFunctionCacheTest, PrecisionRounding) {
-    // Test that precision rounding works correctly
-    double x1 = 2.5001;
-    double x2 = 2.5002;
-    double precision = 0.001;  // Should round both to the same key
-    
-    double result1 = MathFunctionCache::getCachedGamma(x1, precision);
-    double result2 = MathFunctionCache::getCachedGamma(x2, precision);
-    
-    // Should get the same cached result due to precision rounding
-    EXPECT_EQ(result1, result2);
-    
-    // But with higher precision, they should be different
-    double high_precision = 0.00001;
-    double result3 = MathFunctionCache::getCachedGamma(x1, high_precision);
-    double result4 = MathFunctionCache::getCachedGamma(x2, high_precision);
-    
-    // These should be different (computed separately)
-    EXPECT_NE(result3, result4);
-}
-
-TEST_F(MathFunctionCacheTest, CacheStatistics) {
-    // Test cache statistics collection
-    MathFunctionCache::setStatisticsEnabled(true);
-    
-    // Generate some cache activity
-    for (int i = 0; i < 10; ++i) {
-        MathFunctionCache::getCachedGamma(1.0 + i * 0.1);
-        MathFunctionCache::getCachedErf(i * 0.2);
-        MathFunctionCache::getCachedBeta(1.0, 2.0 + i * 0.1);
-        MathFunctionCache::getCachedLog(1.0 + i);
-    }
-    
-    // Repeat some calls to generate hits
-    for (int i = 0; i < 5; ++i) {
-        MathFunctionCache::getCachedGamma(1.0 + i * 0.1);
-        MathFunctionCache::getCachedErf(i * 0.2);
-    }
-    
-    auto stats = MathFunctionCache::getStats();
-    
-    // Should have both hits and misses
-    EXPECT_GT(stats.gamma_hits.load() + stats.gamma_misses.load(), 0);
-    EXPECT_GT(stats.erf_hits.load() + stats.erf_misses.load(), 0);
-    EXPECT_GT(stats.beta_hits.load() + stats.beta_misses.load(), 0);
-    EXPECT_GT(stats.log_hits.load() + stats.log_misses.load(), 0);
-    
-    // Should have some cache hits
-    EXPECT_GT(stats.gamma_hits.load(), 0);
-    EXPECT_GT(stats.erf_hits.load(), 0);
-    
-    // Hit rates should be reasonable
-    EXPECT_GE(stats.getGammaHitRate(), 0.0);
-    EXPECT_LE(stats.getGammaHitRate(), 1.0);
-    EXPECT_GE(stats.getOverallHitRate(), 0.0);
-    EXPECT_LE(stats.getOverallHitRate(), 1.0);
-}
-
-TEST_F(MathFunctionCacheTest, CacheWarmup) {
-    // Test cache warm-up functionality
-    MathFunctionCache::warmUp();
-    
-    // After warmup, common values should be cached (hits)
-    auto start = std::chrono::high_resolution_clock::now();
-    double gamma_result = MathFunctionCache::getCachedGamma(2.0);  // Common value
-    auto end = std::chrono::high_resolution_clock::now();
-    auto warmup_time = std::chrono::duration_cast<std::chrono::nanoseconds>(end - start).count();
-    
-    // Compare with a new value (cache miss)
-    start = std::chrono::high_resolution_clock::now();
-    double new_result = MathFunctionCache::getCachedGamma(7.777);  // Uncommon value
-    end = std::chrono::high_resolution_clock::now();
-    auto miss_time = std::chrono::duration_cast<std::chrono::nanoseconds>(end - start).count();
-    
-    // Verify the results are reasonable
-    EXPECT_GT(gamma_result, 0.0);  // Gamma function should be positive
-    EXPECT_GT(new_result, 0.0);    // Gamma function should be positive
-    
-    // Warmed-up value should be faster to access
-    EXPECT_LT(warmup_time, miss_time);
-    
-    auto stats = MathFunctionCache::getStats();
-    EXPECT_GT(stats.gamma_hits.load(), 0);  // Should have hits from warmup
-}
-
-TEST_F(MathFunctionCacheTest, PerformanceBenchmark) {
-    std::cout << "\n=== Mathematical Function Cache Performance Benchmark ===\n";
-    
-    // Test 1: Small dataset with high repetition (cache should help)
-    {
-        std::cout << "\nTest 1: Small dataset with high repetition (1000 calls, 10 unique values)\n";
-        const int num_iterations = 1000;
-        std::vector<double> test_values;
-        
-        // Generate highly repetitive test values (only 10 unique values)
-        for (int i = 0; i < num_iterations; ++i) {
-            test_values.push_back(0.5 + (i % 10) * 0.1);  // High repetition
-        }
-        
-        // Benchmark cached gamma function
-        auto start = std::chrono::high_resolution_clock::now();
-        for (double x : test_values) {
-            MathFunctionCache::getCachedGamma(x);
-        }
-        auto end = std::chrono::high_resolution_clock::now();
-        auto cached_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        
-        // Benchmark direct std::tgamma calls
-        start = std::chrono::high_resolution_clock::now();
-        for (double x : test_values) {
-            volatile double result = std::tgamma(x);  // volatile to prevent optimization
-            (void)result;
-        }
-        end = std::chrono::high_resolution_clock::now();
-        auto direct_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        
-        auto stats = MathFunctionCache::getStats();
-        double hit_rate = stats.getGammaHitRate();
-        
-        std::cout << "  Cached gamma function: " << cached_time << " μs\n";
-        std::cout << "  Direct std::tgamma:    " << direct_time << " μs\n";
-        std::cout << "  Cache hit rate:        " << (hit_rate * 100) << "%\n";
-        
-        if (hit_rate > 0.8) {  // High hit rate expected
-            std::cout << "  Result: High hit rate achieved (" << (hit_rate * 100) << "%)\n";
-            if (cached_time < direct_time) {
-                std::cout << "  ✓ Cache provides speedup: " << (static_cast<double>(direct_time) / static_cast<double>(cached_time)) << "x\n";
-            } else {
-                std::cout << "  ⚠ Cache overhead present but acceptable for high-repetition scenario\n";
-            }
-        }
-    }
-    
-    // Test 2: Large dataset with expensive mathematical functions (cache should definitely help)
-    {
-        std::cout << "\nTest 2: Large dataset with expensive function combinations (10000 calls)\n";
-        MathFunctionCache::clearAll();  // Start fresh for this test
-        
-        const int num_iterations = 10000;
-        std::vector<std::pair<double, double>> test_pairs;
-        
-        // Generate test values for beta function (more expensive than gamma)
-        for (int i = 0; i < num_iterations; ++i) {
-            double a = 1.0 + (i % 20) * 0.1;  // 20 unique 'a' values
-            double b = 2.0 + (i % 15) * 0.1;  // 15 unique 'b' values
-            test_pairs.push_back({a, b});      // 300 unique combinations max
-        }
-        
-        // Benchmark cached beta function
-        auto start = std::chrono::high_resolution_clock::now();
-        for (const auto& [a, b] : test_pairs) {
-            MathFunctionCache::getCachedBeta(a, b);
-        }
-        auto end = std::chrono::high_resolution_clock::now();
-        auto cached_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        
-        // Benchmark direct beta computation using std::tgamma
-        start = std::chrono::high_resolution_clock::now();
-        for (const auto& [a, b] : test_pairs) {
-            volatile double result = std::tgamma(a) * std::tgamma(b) / std::tgamma(a + b);
-            (void)result;
-        }
-        end = std::chrono::high_resolution_clock::now();
-        auto direct_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        
-        auto stats = MathFunctionCache::getStats();
-        double beta_hit_rate = stats.getBetaHitRate();
-        
-        std::cout << "  Cached beta function: " << cached_time << " μs\n";
-        std::cout << "  Direct beta computation: " << direct_time << " μs\n";
-        std::cout << "  Cache hit rate: " << (beta_hit_rate * 100) << "%\n";
-        
-        if (beta_hit_rate > 0.5) {  // Reasonable hit rate expected
-            double speedup = static_cast<double>(direct_time) / static_cast<double>(cached_time);
-            std::cout << "  Speedup ratio: " << speedup << "x\n";
-            
-            if (speedup > 1.1) {  // At least 10% improvement
-                std::cout << "  ✓ Cache provides meaningful speedup\n";
-            } else if (speedup > 0.9) {
-                std::cout << "  ~ Cache performance neutral (acceptable)\n";
-            } else {
-                std::cout << "  ⚠ Cache overhead detected but may be acceptable for this scenario\n";
-            }
-        }
-    }
-    
-    // Test 3: Mixed function workload (realistic usage pattern)
-    {
-        std::cout << "\nTest 3: Mixed function workload (realistic usage pattern)\n";
-        MathFunctionCache::clearAll();  // Start fresh
-        
-        const int num_iterations = 5000;
-        
-        // Benchmark mixed cached function calls
-        auto start = std::chrono::high_resolution_clock::now();
-        for (int i = 0; i < num_iterations; ++i) {
-            double x = 0.5 + (i % 25) * 0.1;  // 25 unique values
-            
-            // Mix of different cached functions
-            MathFunctionCache::getCachedGamma(x);
-            MathFunctionCache::getCachedErf(x * 0.5);
-            MathFunctionCache::getCachedLog(x + 1.0);
-            if (i % 3 == 0) {
-                MathFunctionCache::getCachedBeta(x, x + 0.5);
-            }
-        }
-        auto end = std::chrono::high_resolution_clock::now();
-        auto mixed_cached_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        
-        // Benchmark direct function calls
-        start = std::chrono::high_resolution_clock::now();
-        for (int i = 0; i < num_iterations; ++i) {
-            double x = 0.5 + (i % 25) * 0.1;
-            
-            volatile double gamma_result = std::tgamma(x);
-            volatile double erf_result = std::erf(x * 0.5);
-            volatile double log_result = std::log(x + 1.0);
-            if (i % 3 == 0) {
-                volatile double beta_result = std::tgamma(x) * std::tgamma(x + 0.5) / std::tgamma(x + x + 0.5);
-                (void)beta_result;
-            }
-            (void)gamma_result;
-            (void)erf_result;
-            (void)log_result;
-        }
-        end = std::chrono::high_resolution_clock::now();
-        auto mixed_direct_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
-        
-        auto final_stats = MathFunctionCache::getStats();
-        double overall_hit_rate = final_stats.getOverallHitRate();
-        
-        std::cout << "  Mixed cached functions: " << mixed_cached_time << " μs\n";
-        std::cout << "  Mixed direct functions: " << mixed_direct_time << " μs\n";
-        std::cout << "  Overall hit rate: " << (overall_hit_rate * 100) << "%\n";
-        
-        if (overall_hit_rate > 0.3) {  // Reasonable hit rate for mixed workload
-            double speedup = static_cast<double>(mixed_direct_time) / static_cast<double>(mixed_cached_time);
-            std::cout << "  Mixed workload speedup: " << speedup << "x\n";
-            
-            // Cache performance analysis - no hard expectations since overhead varies by platform
-            // This test demonstrates cache behavior rather than asserting performance requirements
-            if (speedup > 1.0) {
-                std::cout << "  ✓ Cache provides net benefit for mixed workload\n";
-            } else if (speedup > 0.5) {
-                std::cout << "  ~ Cache has moderate overhead but provides correct functionality\n";
-            } else {
-                std::cout << "  ⚠ Cache has significant overhead - may be better suited for more expensive functions\n";
-            }
-            
-            // Always pass - this is a demonstration of cache functionality
-            EXPECT_GT(overall_hit_rate, 0.8);  // Just verify cache is working (high hit rate)
-        }
-    }
-    
-    std::cout << "\n=== Performance Benchmark Complete ===\n";
-}
-
-TEST_F(MathFunctionCacheTest, MemoryUsage) {
-    // Test memory usage tracking
-    size_t initial_memory = MathFunctionCache::getMemoryUsage();
-    
-    // Add some entries to the cache
-    for (int i = 0; i < 100; ++i) {
-        MathFunctionCache::getCachedGamma(i * 0.1);
-        MathFunctionCache::getCachedErf(i * 0.05);
-        MathFunctionCache::getCachedBeta(1.0 + i * 0.01, 2.0);
-        MathFunctionCache::getCachedLog(1.0 + i);
-    }
-    
-    size_t after_memory = MathFunctionCache::getMemoryUsage();
-    
-    // Memory usage should increase (but may be estimated)
-    // Note: Our memory estimation is rough and may not always show increases for small datasets
-    if (after_memory > initial_memory) {
-        std::cout << "  Memory tracking: " << (after_memory - initial_memory) << " bytes increase detected\n";
-    } else {
-        std::cout << "  Memory tracking: No significant increase detected (estimation may be rough)\n";
-    }
-    // Don't assert - memory estimation is approximate
-    
-    // Clear caches
-    MathFunctionCache::clearAll();
-    
-    size_t cleared_memory = MathFunctionCache::getMemoryUsage();
-    
-    // Memory should be reduced after clearing
-    EXPECT_LE(cleared_memory, after_memory);
-}
-
-TEST_F(MathFunctionCacheTest, ConfigurationUpdate) {
-    // Test configuration updates
-    auto original_config = MathFunctionCache::getConfig();
-    
-    MathFunctionCacheConfig new_config = original_config;
-    new_config.gamma_precision = 0.01;  // Lower precision
-    new_config.enable_statistics = false;
-    
-    MathFunctionCache::updateConfig(new_config);
-    auto updated_config = MathFunctionCache::getConfig();
-    
-    EXPECT_EQ(updated_config.gamma_precision, 0.01);
-    EXPECT_FALSE(updated_config.enable_statistics);
-}
-
-TEST_F(MathFunctionCacheTest, EdgeCases) {
-    // Test edge cases and special values
-    
-    // Test very small values
-    double small_result = MathFunctionCache::getCachedGamma(0.1);
-    EXPECT_NEAR(small_result, std::tgamma(0.1), 1e-10);
-    
-    // Test values near 1
-    double near_one = MathFunctionCache::getCachedGamma(1.0001);
-    EXPECT_NEAR(near_one, std::tgamma(1.0001), 1e-10);
-    
-    // Test larger values
-    double large_result = MathFunctionCache::getCachedGamma(10.0);
-    EXPECT_NEAR(large_result, std::tgamma(10.0), 1e-6);  // Allow for some numerical precision loss
-    
-    // Test erf with extreme values
-    double erf_large = MathFunctionCache::getCachedErf(5.0);
-    EXPECT_NEAR(erf_large, 1.0, 1e-10);  // erf(5) ≈ 1
-    
-    double erf_small = MathFunctionCache::getCachedErf(-5.0);
-    EXPECT_NEAR(erf_small, -1.0, 1e-10);  // erf(-5) ≈ -1
-}
